Add files via upload
Major update:

Created and added gui, modified main generator file to allow for custom selection, as of current the only nonfunctioning gui utilities are the custom attribute generation, and show-in-folder feature.  

added a few more options for generation.

did some cleaning of the code to remove redundant or debugging features.